Move Typing Speed Test to Beginner_Projects folder
